+++ tests/__mocks__/linter-provider-eslint.ts
