+++ tests/jest-plugin-fs.d.ts
@@ -1,0 +1,1 @@
+declare module "jest-plugin-fs";
